Add custom hook with audio settings
