+++ src/components/AudioButton/AudioButton.md
@@ -5,5 +5,8 @@ Audio Button
   songUrl={
     'https://p.scdn.co/mp3-preview/3011a4d543fca575e3f89bdb80f80c559161a96c?cid=5f7fa324ec124c5da5672dd5d6fa9b2f'
   }
+  id={'nj8se0eqez237'}
+  currentSong={'nj8se0eqez237'}
+  setCurrentSong={() => {}}
 />
 ```
